// src/shared/enums/screens.enum.ts
export enum ScreensEnum {
  HOME_SCREEN = 'home-screen',
  SETTINGS_SCREEN = 'settings-screen',
  HISTORY_SCREEN = 'history-screen',
  PLAYER_AMOUNT_SCREEN = 'player-amount-screen',
  PLAYER_NAMES_SCREEN = 'player-names-screen',
  PLAYER_SCORES_SCREEN = 'player-scores-screen',
  PLAYER_BIDDING_SCREEN = 'player-bidding-screen',
  PLAYER_BIDDING_RESULTS_SCREEN = 'player-bidding-results-screen',
  PLAYER_RANKING_SCREEN = 'player-ranking-screen',
}